Added IMT to code

+++ Afrinet.Models/Afrinet/Transaction.cs
@@ -2,8 +2,8 @@
 namespace Afrinet.Models;
 public class Transaction
 {
- 
-        public string? Id { get; set; }
+
+    public string? Id { get; set; }
     public string? InitatingAccountID { get; set; }
 
     public string? SendingAccountID { get; set; }
@@ -28,7 +28,7 @@ public class Transaction
 
     public string? ReversedTransactionID { get; set; }
 
-    public long TotalFees { get; set; }
+    public double TotalFees { get; set; }
 
     public long TotalTax { get; set; }
 
@@ -36,10 +36,10 @@ public class Transaction
 
     public List<Tax> Taxes { get; set; } = new List<Tax>();
 
-    public string? Channel {get; set;}
-    public long Amount {get; set;}
+    public string? Channel { get; set; }
+    public double Amount { get; set; }
 
-    public Voucher? Voucher {get; set;}
+    public Voucher? Voucher { get; set; }
